refactor(response): unify API response types and introduce SingleItemResponse for consistent data handling
- Rename DataResponse to CollectionResponse and update to extend ApiResponse<T[]>
- Update PageResponse to extend CollectionResponse and fix references
- Refine ApiResponse to use generic data type T (not T[])
- Add SingleItemResponse interface for single resource responses
- Update SubscriberApi import paths for PageResponse
- Update documentation and structure to reflect new response types
