+++ client/apps/web/src/shared/response/PageResponse.ts
@@ -1,15 +1,15 @@
 /**
- * Represents a paginated response that extends {@link DataResponse} with cursor-based pagination.
+ * Represents a paginated response that extends {@link CollectionResponse} with cursor-based pagination.
  *
  * @template T - The type of the data contained in the response.
- * @extends DataResponse<T>
+ * @extends CollectionResponse<T>
  *
  * @property {string | undefined | null} prevPageCursor - The cursor for the previous page, if available.
  * @property {string | undefined | null} nextPageCursor - The cursor for the next page, if available.
  */
-import type DataResponse from "./DataResponse";
+import type CollectionResponse from "./CollectionResponse.ts";
 
-export default interface PageResponse<T> extends DataResponse<T> {
+export default interface PageResponse<T> extends CollectionResponse<T> {
 	readonly prevPageCursor: string | undefined | null;
 	readonly nextPageCursor: string | undefined | null;
 }
